requirements.txt
numpy>=1.18.5
opencv-python>=4.5.1
streamlit>=1.12.0
ultralytics>=8.0.0
Pillow>=9.0.0
torch>=1.8.0